Add files via upload
As we all know that Olympics is the most prestigious Sporting Event lets have a look at the history of the Olympic games to get more insights of it.
This Dashboard comprises of all the Events played from 1896 Olympic Games to 2016 Olympics Games featuring all Male/Female Athletes from all over the world which has featured in the Olympic Games.
This dashboard contains which Athletes won Gold, Silver or Bronze Medals for country and which are the top Athletes which won the most Medals(Gold, Silver or Bronze) in the history of the Games and which countries are leading the Medal tally in Summer/Winter Olympic Games. Furthermore, this Dashboard also covers total count of medals according diffferent sports.